src/requirements/development.txt
isort
pylama
pytest
pytest-cov
pytest-django
pytest-sugar
ipython
freezegun
